requirements.txt
streamlit==1.28.1
pandas==2.0.3
plotly==5.17.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
numpy==1.24.3
scipy==1.11.4
boto3==1.34.0
python-dotenv==1.0.0
requests==2.31.0